Fix this YUGE bug due to Giants bad design.
The currentDay of vanilla environment is only updated on an MP client
after the first :update call. That means that everything we do
inside loadMap and readStream on a client uses the wrong currentDay.
This results in wrong growth transitions, wrong seasons, wrong temps,
wrong daylight, wrong animals. WRONG EVERYTHING.
'Fixed' it by updating the currentDay ourselves in the
ssEnvironment:readStream