Fix bug causing other mods to not be able to append to start and stopMotor.
The issue is that we do not call the superFunc for our overwritten function:
we don't want to call the original function. However, that means any
overwrittenFunction or appendedFunction from other mods are never called.
To solve this we replace the Motorized methods instead. This is not good
practice but we copied Giants code and changed a bunch of it so I guess
it should work just fine.