Refactor RV again for advanced algo.
Use params to give the update functions info about what to show.
Store the latest visuals (not console) in the game and send it to multiplayer.
Update code accordingly.

Now, the code for getVisualSeason can use the latestVisuals to prevent from
going from spring to bare trees, for example.

Now, visuals are possibly updated every night