Fix RV bug for placeable trees.
The default materials were only loaded on loadMap. This is an
issue when any of the placeable trees were not already on
the map (in GE).
The fix was to load default materials from placeables when
they are placed.
This is only done if not done before for given placeable.