Lowered the brightnis
I made the brightnis lower ( like the V1 of Seasons) because i think it looks a lot better this way. The night stays the same just the day is affected. 
The brightnis as an example 42-47 is taken from the V 1 of Seasons so nothing realy new just a combination from V1 and V1.1