Fix major issue with animals in mp.
The game sent animal info to the client before values were adjusted there. The client looks if there is a positive milk production on the animal, and if there is, reads the fill level. The values were not in sync, so the reading and writing was not either, causing major synchronization issues